Interface 분리 작업
